+++ crates/volta-core/src/inventory/serial.rs
@@ -100,51 +100,6 @@ impl PackageCollection {
     }
 }
 
-#[derive(Serialize, Deserialize, Debug)]
-pub struct YarnIndex(Vec<YarnEntry>);
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct YarnEntry {
-    /// Yarn releases are given a tag name of the form "v$version" where $version
-    /// is the release's version string.
-    #[serde(with = "version_serde")]
-    pub tag_name: Version,
-
-    /// The GitHub API provides a list of assets. Some Yarn releases don't include
-    /// a tarball, so we don't support them and remove them from the set of available
-    /// Yarn versions.
-    pub assets: Vec<YarnAsset>,
-}
-
-impl YarnEntry {
-    /// Is this entry a full release, i.e., does this entry's asset list include a
-    /// proper release tarball?
-    fn is_full_release(&self) -> bool {
-        let release_filename = &format!("yarn-v{}.tar.gz", self.tag_name)[..];
-        self.assets
-            .iter()
-            .any(|&YarnAsset { ref name }| name == release_filename)
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct YarnAsset {
-    /// The filename of an asset included in a Yarn GitHub release.
-    pub name: String,
-}
-
-impl YarnIndex {
-    pub fn into_index(self) -> Fallible<super::YarnIndex> {
-        let mut entries = BTreeSet::new();
-        for entry in self.0 {
-            if entry.is_full_release() {
-                entries.insert(entry.tag_name);
-            }
-        }
-        Ok(super::YarnIndex { entries })
-    }
-}
-
 // see npm registry API doc:
 // https://github.com/npm/registry/blob/master/docs/REGISTRY-API.md
