+++ crates/volta-core/src/resolve/serial.rs
@@ -1,4 +1,4 @@
-use std::collections::HashSet;
+use std::collections::{BTreeSet, HashSet};
 use std::iter::FromIterator;
 
 use crate::version::{option_version_serde, version_serde};
@@ -48,5 +48,50 @@ where
     match String::deserialize(deserializer) {
         Ok(_t) => Ok(true),
         Err(_e) => Ok(false),
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct RawYarnIndex(Vec<RawYarnEntry>);
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct RawYarnEntry {
+    /// Yarn releases are given a tag name of the form "v$version" where $version
+    /// is the release's version string.
+    #[serde(with = "version_serde")]
+    pub tag_name: Version,
+
+    /// The GitHub API provides a list of assets. Some Yarn releases don't include
+    /// a tarball, so we don't support them and remove them from the set of available
+    /// Yarn versions.
+    pub assets: Vec<RawYarnAsset>,
+}
+
+impl RawYarnEntry {
+    /// Is this entry a full release, i.e., does this entry's asset list include a
+    /// proper release tarball?
+    fn is_full_release(&self) -> bool {
+        let release_filename = &format!("yarn-v{}.tar.gz", self.tag_name)[..];
+        self.assets
+            .iter()
+            .any(|&RawYarnAsset { ref name }| name == release_filename)
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct RawYarnAsset {
+    /// The filename of an asset included in a Yarn GitHub release.
+    pub name: String,
+}
+
+impl RawYarnIndex {
+    pub fn into_index(self) -> Fallible<super::yarn::YarnIndex> {
+        let mut entries = BTreeSet::new();
+        for entry in self.0 {
+            if entry.is_full_release() {
+                entries.insert(entry.tag_name);
+            }
+        }
+        Ok(super::yarn::YarnIndex { entries })
     }
 }
